requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
scikit-learn
streamlit
rich
faker
PyMuPDF
beautifulsoup4
python-dateutil
scipy
sentry-sdk
tqdm
